Implemented Server side Transactions
Issue #139
- new begin/commit/rollback commands in HTTP API
- new begin/commit/rollback methods in RemoteDatabase
- server side transactions are implemented as sessions that start at begin() command
- sessions expires with a configurable timeout (=30 secs)

+++ postgresw/src/test/java/com/arcadedb/postgres/BaseGraphServerTest.java
@@ -242,7 +242,7 @@ public abstract class BaseGraphServerTest {
   protected void deleteDatabaseFolders() {
     if (databases != null)
       for (int i = 0; i < databases.length; ++i) {
-        if (databases[i] != null)
+        if (databases[i] != null && databases[i].isOpen())
           databases[i].drop();
       }
 
@@ -255,7 +255,7 @@ public abstract class BaseGraphServerTest {
     Assertions.assertTrue(DatabaseFactory.getActiveDatabaseInstances().isEmpty(), "Found active databases: " + DatabaseFactory.getActiveDatabaseInstances());
 
     for (int i = 0; i < getServerCount(); ++i)
-      FileUtils.deleteRecursively(new File(getDatabasePath(i)));
+      FileUtils.deleteRecursively(new File(GlobalConfiguration.SERVER_DATABASE_DIRECTORY.getValueAsString() + i + "/"));
     FileUtils.deleteRecursively(new File(GlobalConfiguration.SERVER_ROOT_PATH.getValueAsString() + "/replication"));
   }
